Ajudando Freitas part 4

+++ src/main/java/com/github/hortykut/hortykut/controller/ProdutoController.java
@@ -1,9 +1,15 @@
 package com.github.hortykut.hortykut.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
 import com.github.hortykut.hortykut.model.Produto;
+import com.github.hortykut.hortykut.repository.CategoriaRepository;
 import com.github.hortykut.hortykut.repository.ProdutoRepository;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -11,41 +17,57 @@ import java.util.Optional;
 @RequestMapping("/produto")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class ProdutoController {
-	
-    @Autowired
-    private ProdutoRepository produtoRepository ;
-    
-    @GetMapping
-    public List<Produto> listarProdutos() {
-        return produtoRepository.findAll();
-    }
-    
-    @GetMapping("/{id}")
-    public Optional<Produto> buscarProdutoPorId(@PathVariable Long id) {
-      return produtoRepository.findById(id);
-    }
-    @PostMapping
-    public Produto criarProduto(@RequestBody Produto produto){
-        return produtoRepository.save(produto);
-    }
-    @PutMapping
-    public Produto atualizarProduto(@RequestBody Produto produtoAtualizado) {
-        Optional<Produto> produtoExistente = produtoRepository.findById(produtoAtualizado.getId());
 
-        if (produtoExistente.isPresent()) {
-        	Produto produto = produtoExistente.get();
-            produto.setProduto(produtoAtualizado.getProduto());
-            produto.setResponsavel(produtoAtualizado.getResponsavel());
-            produto.setValor(produtoAtualizado.getValor());
-		 produto.setDescricao(produtoAtualizado.getDescricao());
-		produto.setFoto(produtoAtualizado.getFoto());
-            return produtoRepository.save(produto);
-        } else {
-            throw new IllegalArgumentException("ID do usuário inválido: " + produtoAtualizado.getId());
-        }
+    @Autowired
+    private ProdutoRepository produtoRepository;
+
+    @Autowired
+    private CategoriaRepository categoriaRepository;
+
+    @GetMapping
+    public ResponseEntity<List<Produto>> getAll() {
+        return ResponseEntity.ok(produtoRepository.findAll());
     }
-    @DeleteMapping("/{id}")
-    public void deletarProduto(@PathVariable Long id) {
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Produto> getById(@PathVariable Long id) {
+        return produtoRepository.findById(id).map(resposta -> ResponseEntity.ok(resposta))
+                .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+    }
+
+    @GetMapping("/titulo/{titulo}")
+    public ResponseEntity<List<Produto>> findAllByTitulo(@PathVariable String titulo) {
+        return ResponseEntity.ok(produtoRepository.findAllByTituloContainingIgnoreCase(titulo));
+    }
+
+    @PostMapping
+    public ResponseEntity<Produto> post(@Valid @RequestBody Produto produto) {
+        if (categoriaRepository.existsById(produto.getCategoria().getId()))
+            return ResponseEntity.status(HttpStatus.CREATED).body(produtoRepository.save(produto));
+
+        throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Categoria não existe!", null);
+    }
+
+    @PutMapping
+    public ResponseEntity<Produto> put(@Valid @RequestBody Produto produto) {
+        if (produtoRepository.existsById(produto.getId())) {
+            if (categoriaRepository.existsById(produto.getCategoria().getId())) {
+                return produtoRepository.findById(produto.getId())
+                        .map(resp -> ResponseEntity.status(HttpStatus.OK).body(produtoRepository.save(produto)))
+                        .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+            }
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Categoria não existe!");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    }
+
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @DeleteMapping("{id}")
+    public void delete(@PathVariable Long id) {
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        if (produto.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         produtoRepository.deleteById(id);
     }
 }
